Add dragging and delete functionality to mobile
This update should add functionality to mobile
devices like on PC, for dragging and deleting
a note. Instead of click holding to drag and right
clicking to delete, double taps should delete,
and touch holding should drag. This will be
updated again if corrections are needed from
issues yet to be seen.